types: Name and document the "this default is unreachable" pattern.
I first started using this pattern in de76ca2f7, and explained
it slightly more in a later commit message 93c7a4fc9... but it's
pretty obscure when looking at the actual code in the tree.
Wrap it in a name which gives some idea of the point, and also
write detailed jsdoc for it.