msglist js: Give arrayFrom a real type -- and fix a type error.
Flow thinks `Array.prototype.slice` returns an array of `empty`,
i.e. of magical shapeshifter values that can be whatever you
need them to be.

Teach it, in our usage at least, that the element type of the array
is just the element type of the array-like object we passed in.

And in fact we turn out to have had a type error that this was
hiding!  Now that the elements in `handleUpdateEventMessagesRead`
are understood to be `HTMLElement`s rather than magical
shape-shifters, Flow knows what types `setAttribute` is supposed
to accept, and that the attribute value should be a string.
So, fix that.