time pills: Convert most uses of rem to px.
This is a pure refactor, in the default case where the root font size
is 16px; we convert the values exactly for that case, and in the next
commit we'll adjust them to values that align nicely to our grid.

When the font size is expanded, this makes a welcome change:
previously the larger font would cause the pills' padding and inset
margin to grow, encroaching on the content precisely when it's likely
to be crowded already.  Now it doesn't.