persist-migrate: Fix to actually work again, oops.
Fixes #3553.

Looks like I broke this in 397d427ca.  The effect was that migrations
wouldn't do anything; the only thing the migrations system would do is
to update the `migrations` metadata *as if* the needed migrations had
been run.

We haven't added any migrations since then (trying to do so is how I
discovered this bug today).  The buggy commit went out in the 25.0.114
release; there were no new migrations in that release, and the
previous one was in 24.0.113, adjusting `realm.emoji`.

So the total effect is that anyone upgrading directly to 25.0.114 (or
later) from a release prior to 24.0.113 wouldn't get the migrations in
between.  That isn't common, but for the users it did affect, it could
cause a crash.  In particular, for anyone upgrading from before
22.0.107 (almost 5 months earlier), they'd miss a migration affecting
`accounts`, which unlike most of our data doesn't get regularly
clobbered by a `/register` response.

Well, fix the bug.  In an upcoming commit we'll follow up with a
remedial migration, too, to fix any affected installs that are still
using the app.

A shame this file isn't tested. :-/  That's one root cause of this bug
getting through.