hidden email: Smooth transitions with a bit of fake blank data.
We're about to start using this `email` property for `getOwnEmail`;
and we'd like to have it throw an error if we don't actually have a
valid own-email because there isn't a logged-in active account.

Unfortunately it turns out -- as we learn from trying that -- that
there are existing parts of our code that can end up trying to get at
the own-email when there is no valid answer.  That's a bug, but for
now just keep it papered over by making up some blank data (like we do
in a lot of our legacy code), and make a TODO comment.

The harder-line `initialState` value still has a foothold, in that we
use it on first startup.