AccountDetails: Show name instead of email.
This brings this UI in line with the corresponding UI in the webapp,
which is the card that pops up on clicking on a user's name or avatar.

The user's email address can be helpful information, but their name is
usually more helpful.  Also, under an org setting that's currently in
beta after being a frequent request for a long time, users' real email
addresses won't even be available to normal users -- so the `email`
property will be an opaque address made up by the server.  (See #3196.)
So we certainly don't want to show that here.

While we're here, also swap the order of the name and the status
indicator, again following the webapp.