webview updates: Don't scan the `read` flags twice.
The deep comparison in `isEqual` has to scan the entire objects on
both sides, if they're not identical (with ===) but turn out to have
equal contents, or nearly so.

In the parent commit, that means that if just a few out of a large set
of messages have become read -- the common case! -- then we might scan
nearly the entire data structures in `isEqual` to learn that, only to
do so again in `updateRead` to compute the list of changes.

Instead, just compute the list of changes.  If it's empty, we can
still skip sending it, and we've only done the same work we'd have
ended up doing in `isEqual`.