webview updates: Cut an illogical check; fix ill-typed tests.
This condition makes no sense if taken seriously: if it were
actually possible for these `flags` objects to be missing, that'd
surely be a way of saying they were empty... and then suppressing a
"read" update when the *previous* state had nothing read would be
completely backward.

Fortunately this isn't that bug -- in fact, the `flags` property
is required in the type, so this is always an object and the
condition does nothing.

Except in some of our tests, which are flagrantly ill-typed.  Fix
them far enough to not break on this code, using our well-typed
`exampleData` fixtures to get an appropriate `flags` object.