eslint: Disable various upcoming rules we don't want.
We're about to upgrade our ESLint major version to 5, and all
ESLint-related packages to the latest compatible with ESLint 5
(i.e. in most cases the very latest.)

Those changes will introduce, enable, or make more demanding a
number of rules; some of them useful, many of them not.  Here,
disable most of the rules that aren't.

The next few commits will make further preparations for the upgrade.