yarn: Remove `react-deep-force-update` noise from package.json.
This `resolutions` entry was added about 15 months ago in 41d60ff85.
As that commit message (and the GH comment it links to) explain, it
was a way of forcing Yarn to upgrade this package -- an indirect
dependency -- to 1.1.1, to pick up a bugfix.

Rereading that GH comment that I relied on in making that change, it
actually recommended making the `resolutions` entry temporary, leaving
it in only long enough to run `yarn` in order to update `yarn.lock`.

Now, a recent `yarn upgrade` picked up the first update to that
package since then, to 1.1.2; and for some reason also caused the
`resolutions` entry to get copied to `dependencies`, where it's noise.

Cut it there, and cut also the `resolutions` entry as originally
recommended.  Then run `yarn` to clear out the corresponding entry in
`yarn.lock`; and now a `yarn upgrade` leaves this all untouched.