webview: Fix types thanks to new libdef.
Now that there's a real libdef for the WebView component (since the
parent commit), Flow can find type errors.

First, use the real WebView type for the reference to the component
that we keep around; and fix the type of a callback, using read-only
properties.

Then there's a bit where we're genuinely doing something the stated
types don't allow.  That's because the method we're calling, though
really implemented, isn't documented!  I just filed an issue for that
(though we learned about it the other day); link to that in a fixme.