emoji/data: Rewrite false `reduce`s to plain `forEach`.
Not the most beautiful possible way to write these -- but this is a
maximally boring, mechanical way to rewrite them, which we can spruce
up in the next commit.  And it does get rid of the confusing mutating
`reduce`, as well as put the starting value at the start of the code.