flow: Remove an old workaround that causes its own trouble.
With this `[empty]: mixed` in place, a snippet like this:

  (emoji: ImageEmojiType) => {
    ({ ...emoji, xyz: 123 }: ImageEmojiType);
  };

is accepted, even though in fact it's ill-typed.  Dang.

We long ago sorted out the specific confusion that motivated putting
in this workaround; and since then I don't think it's helped us spot
any others.  So, go back to the natural way of writing the type.

Ultimately the right solution to the risk of confusing objects-as-maps
and objects-as-structs is that we should use `Map` for maps, instead
of objects-as-maps.