flow: Add dummy libdef for react-native-document-picker.
In e17b1a1ee we replaced the stub libdef for this library with
our own real one.  But `yarn update-libdefs` (and the underlying
`flow-typed update`) doesn't notice our real one, and keeps wanting
to generate a new stub.

It's annoying to have to keep deleting the stub whenever we use
`yarn update-libdefs`, so just let it have its way.  Fortunately
the real libdef takes precedence, empirically: if we artificially
introduce a type error in ComposeMenu.js where this library is
used, Flow indeed catches it.

[greg: added context]