kotlin: Simplify `String#split` usage in NotificationHelper.
This odd-looking usage was the legacy of the automated conversion to
Kotlin from Java in 7d96c7f7d; for discussion see 9fd7be59f.

Fortunately we aren't affected by the surprising semantics of Java's
`String#split`, which this emulates; so just simplify it to a plain
`split` call with the natural semantics that has in Kotlin.