api: Use `import *` to clean up importing API bindings.
This follows up on b8010deca.  By saying `import * as api`
rather than `import api`, we avoid needing to stick these
functions as properties on a default export and can simply
export them directly.  This lessens the repetition in the
api tree's index.js .  We also get to avoid having to disable
a lint rule.

With those improvements, this is now clearly cleaner than the
alternative, so go and switch to this style everywhere.
Done automatically with:

  $ perl -i -0pe '
      if (s<import (?:api|(?:api, )?\{ ([\w, ]+) \}) from (.*/api.);>
           <import * as api from $2;>) {
          my @names = split /, /, $1;
          foreach my $name (@names) {
              s/\b$name\b/api.$name/g
          }
      }
    ' src/**/*.js

  $ perl -i -ne '
      print unless (m/no-named-as-default-member/)
    ' src/**/*.js

Then deleted the default-export in src/api/index.js .