android notif: Properly group different senders' messages.
Fixes #3591.

Previously this code would treat messages from different senders as
belonging to different "conversations", even if they were in a single
group-PM thread or a single stream and topic.  Fix that.

As a result, we might for example display a notification like this:

  5 messages in 2 conversations
  name 1, name 2: message 2
  name 3, name 4, name 5: message 5

which previously would have looked like this:

  5 messages in 5 conversations
  name 1: message 1
  name 2: message 2
  name 3: message 3
  name 4: message 4
  name 5: message 5

[greg: added `numConversations` logic to fix "1 conversations" in
 title and summary; stuck with `sender.email` in `buildKeyString`,
 for until we explicitly desupport pre-1.8 Zulip servers; formatted
 code; squashed.]