android notif: Show a given conversation's senders in order.
This doesn't handle everything that'd be desirable -- e.g. if the
senders are either A B A or A B B, we'll show "A, B:" and the last
message either way, and it's impossible for the user to tell if the
message was from A or from B.

But at least it's deterministic, and in particular in the simple case
of messages from A and then from B it consistently shows "A, B:".