url: adjust hasProtocol regex
The regex in hasProtocol was subtly flawed: it would recognize
technically-legal protocol-less URLs of the form `example.org/http://`
as having a valid protocol.

This replacement regex regects that string. For consistency with
`fixRealmUrl`, it also ignores leading spaces.

In preparation for a future commit, it also carefully specifies
appropriate capturing groups. New tests will be added in that future
commit.