tools/test: Check Node version >= 10 before trying to run Jest.
If you're using Node 8.x, the previous LTS version, then 338036e0c
causes every test that would import `ComposeMenu.js` (which is a large
fraction of them) to break with a loud syntax error, as described in
the parent commit.

Our setup instructions in build-run.md say to use Node 10.x, but
that's pretty easy to miss, especially if you already have Node lying
around; and the syntax errors are not very helpful for diagnosing the
problem.  (See e.g. #3619.)  So, make our own quick version check first.

(With Node 11.x there's a *different* issue described in the
troubleshooting section of build-run.md .  But that comes before this
step anyway, so there's nothing to be gained by enforcing an upper
bound here; may as well leave it open for when Node versions >= 11 do
eventually work.)

I'm actually puzzled why the Node version matters here, because I'd
have thought it's Babel's job to compile the fancy syntax out.  As the
commit message on 338036e0c says, our Babel config appears to have the
plugin enabled that would compile this syntax out.  I suspect there's
something wrong in our Jest config for Babel, but haven't dug into it.
In any case it seems for the best to standardize on a version of Node.