Icons: Extend type-checking so it checks the icon names, too.
Before this change, if we write a `name` prop that mistypes an icon's
name, or refers to an icon that for any reason doesn't exist, Flow
won't give an error.

Since cd899bbe0 we've had checking of this at call sites of
`NavButton`, which explicitly takes a prop of type `IconNames`.  This
change adds such type-checking for call sites of `Icon` itself, and at
the many references to specific icons in this file.