yarn: Bring yarn.lock to fixpoint on eslint-utils.
In 5c0b3a09a we upgraded an indirect dependency in yarn.lock by adding
a "resolutions" object temporarily to package.json , then running
`yarn`, then taking the "resolutions" object back out.

That leaves yarn.lock in a slightly unstable state: it still has a
memory of the "resolutions" constraint itself, and when we run `yarn`
again that gets taken back out.  So, do that and commit the result.