search: Move request-ID allocation out of inner helper.
This is a pure refactor.  But if handleQueryChangeInner were wrapped
in a throttle or debounce in the most obvious way, there'd be a bug
which this fixes.  The possibility of doing that is the main reason
this function is separate, so seems best to clean that up while it's
fresh in our minds.