presence + typing-status: Use consistent "now" times.
Not really any harder to do, and an essential part of writing
time-related code in a way that can be reasoned about.

If we're extremely starved for CPU, so that a material amount of time
passes between the two `new Date()` calls in each function in the old
version of this code, then the old code would throttle these requests
more than intended, and this change fixes that.  Outside of those
conditions, this change is NFC for practical purposes.