android notif: On open when app in background, emit the event.
When the user has opened the app and then moved on to something else,
it may still be in memory for a while, including the JS environment.
In this state, if the user opens a notification and we simply launch
the app, the startup logic that looks at `initialNotification` won't
run.  This state corresponds to a `lifecycleState` of BEFORE_RESUME,
and an `appStatus` of BACKGROUND.

We do need to launch the app, to get it into the foreground; but we
also need to go and emit the event, so that the already-initialized
JS code sees it and navigates to the right conversation.

Fixes: #3582
Reported-by: Vishwesh Jainkuniya <gitvishwesh@gmail.com>
Debugged-by: Vishwesh Jainkuniya <gitvishwesh@gmail.com>