exampleData: Use deepFreeze everywhere.
We've had a practice of doing this in our legacy tests, but I seem to
have neglected to transfer it over in most of the conversions to this
form.  It's easy to do and potentially catches bugs of a class that
would be nasty to debug, so go and do it throughout.