yarn: Make it possible to use `yarn link` to do work on dependencies.
This is an extremely handy feature of Yarn for doing development on a
dependency in tandem with the app itself.  The new `@zulip/shared` is
a prime use case for this feature, though it should also be useful for
third-party dependencies.

Here, take care of a few ways that our toolchain -- more precisely,
three different Facebook tools in our toolchain -- doesn't out of the
box handle `yarn link` well.  Also document it.

Awkwardly, two of the three (Flow and Metro) require explicitly
writing down a list of the dependencies that might turn out to be
behind a `yarn link`.  For now, just list `@zulip/shared`.  It's easy
to add more as they come up, imperfect though that workflow is.

---

An alternative workaround-hack I experimented was: make the practice
be that if you use `yarn link`, you also make sure to run `yarn` in
the linked-to tree.  With that solution, the Jest config bit could be
eliminated, and the Metro `extraNodeModules` bit.  The Flow bit and
the Metro `watchFolders` bit were still needed.

What caused me to abandon that hack and go for this one is that if a
module in the linked-to package has an ES6 `import` statement, then
under our Babel config it will end up requiring `@babel/runtime`...
so that hack only works if the package's own package.json declares
that as a dependency.  Which it quite reasonably might not -- it's
really *our* Babel config that causes that module to end up requiring
the Babel runtime.

(Also that hack was more appealing, relatively, before I added the
`packageDeps` logic -- when this alternative would have meant
duplicating here the list of dependencies of `@zulip/shared`.)