jest: _actually_ set up jest-extended
The preceding commit to set up `jest-extended` lost its configuration
in the shuffle from one config file to another, and the example of use
(which was intended to serve as a test) actually used functions already
present in the current version of Jest.

Properly configure `jest-extended`, and use a function known to be
provided by it.