connection-status banner: Remove animation
When disconnecting from a VPN on Android 9 and later, the connection
state goes from 'vpn' to 'none' to 'wifi' very quickly. This can
result in the online banner getting stuck in mid-animation, until
some other event (e.g. the user scrolling) causes a redraw.

Since the animation is nonessential here (and this is hopefully a
rarely-seen UI component anyway), for now, just make it static.

Effectively reverts 90f66db91.