rewriteImageUrls [nfc]: Use Element.getAttribute().
This is a bit simpler than going through `attributes`... and as a
bonus, Flow knows the right type for it.

Confirmed empirically (in a Chrome 37 devtools console) that this
gets the same, raw, result.  Which is good because that's also
clearly the right answer architecturally: the getter for `.src`
may be magical, but both `.getAttribute()` and `.attributes` are
generic interfaces that shouldn't know about any special quirks
for particular attributes.