rewriteImageUrls [nfc]: Make a pass at "why" vs. "what" comments.
In general, if a comment would just express in prose the same thoughts
as the code itself and at similar length, it's preferable to let the
code speak for itself.  That lets the reader focus on the code, as the
authoritative source on what the code actually does.

On the other hand, some thoughts can be much harder to express clearly
in code than in comments around it: the reason something's done a
certain way; a quirk or pitfall in some outside code being used; the
explanation of a nontrivial algorithm and why it works; perhaps most
ubiquitously, the intended meaning, contract, or invariants of
something that's exposed as some form of API.

So another benefit of leaving out "what" comments that don't add as
much information to the code is that it leaves more room, physically
and mentally, for the more valuable "why" comments.

Here, make an editing pass over this function to add several more
"why" comments and delete or shorten "what" comments.

Some of the old "what" comments touched on the function's intended
behavior and interface, which is good "why" material.  But we already
have that nicely expressed in the function's doc comment, so we can
leave the details for there.  What the comment can add is the
relationship between the spec and this particular part of the code:
e.g., "Find the image elements to act on" tells the reader this is
supposed to be the complete list of elements described in the spec,
and not just one of several waves of them.