lint/flow: apply $ReadOnly to connect()-produced components' props
Flow's `Diff<>` utility type is bugged: it doesn't preserve the
variance of properties. [1]

This means, in turn, that our type harness around `connect()` can't
preserve whether the individual members of a given prop-type are
immutable. Fortunately, we don't have to actually _preserve_ that; we
can just blindly make everything `$ReadOnly`. (React requires that
props be immutable anyway, and a following commit in this patch-set
will add a lint that helps ensure it.)

(This comes early in the patch set -- and indeed, was only noticed at
all -- due to some subtyping expectations in titleButtonFromNarrow.js.)

[1] https://github.com/facebook/flow/issues/6225