webview js: Get iOS-vs.-Android flag from RN, rather than userAgent.
It turns out that on certain newer models of iPad, the user-agent
string has reached new heights of the absurdity that user-agent
strings are notorious for.  In particular, on a latest-model iPad
(just "iPad" with no further modifiers), it's this:

  Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15) AppleWebKit/605.1.15 (KHTML, like Gecko)

It probably really is that version of Apple's WebKit, but it isn't
much else of that; certainly not an Intel Mac.  Most relevant to us,
it doesn't mention "iPad" or anything like it -- with the result that
we don't realize it's iOS, and we end up attaching our `message`
listener to the `document` object rather than `window`, causing
everything the outer RN environment tries to send it to vanish into
the ether.  (That's the only place we use this flag.)

Rather than add another wrinkle to an attempt to make sense of
user-agent strings, just have the React-side code inject this
information from its own notion of which platform we're on.

Fixes: #3657
Signed-off-by: Greg Price <greg@zulipchat.com>
Signed-off-by: Ray Kraesig <rkraesig@zulipchat.com>