webview build [nfc]: Build straight into destination folder.
This "staging" step was intended as a way to help the developer
hand-inspect the output for stale files.  Better to prevent them
automatically, by continuing to use tools like `rsync --delete`.
Meanwhile, the extra indirection complicates the code and
complicates reasoning about it.  So, simplify it away.

This change actually has no effect on the current code's behavior,
because we're having a source directory double as the "staging"
directory, and that's the *only* source directory involved in this
build.  The real effect comes when we add another source directory:
we'll copy from that one straight into the destination, rather than
first into this source directory and then again to the destination.

This change drops a check that `$root/src/webview/static/` exists.
That check is pretty redundant already -- the path is within our own
tree -- but in any case `rsync` will give an error if the directory
it's supposed to copy from is missing.