sentry: add tests to verify assumed behavior
The code in sentry.js makes some currently-unsupported statements
about how Sentry works when uninitialized, and the code in logging.js
relies on them. These tests act as the needed support.

They also incidentally act as a minor check on the handwritten Flow
type definitions, ensuring at least that the relevant functions exist
and are callable with the given values -- which was not the case in an
earlier revision!